requirements/base.txt
asgiref==3.4.1
certifi==2020.11.8
chardet==4.0.0
Django==3.2.8
Pillow==8.3.0
psycopg2-binary==2.9.1

django-admin-honeypot==1.1.0
django-ckeditor==6.0.0
django-extensions==3.1.3
django-js-asset==1.2.2
django-meta==2.0.0
django-modeltranslation==0.17.3
django-mptt==0.11.0
django-newsletter==0.9.1
django-pwa==1.0.10
django-recaptcha==2.0.6
django-taggit==1.5.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
